requirements.txt
pythonwhois
dnspython
pygeoip
shodan
bs4
lxml

